requirements.txt
# Customer Support Agent — Python Dependencies
# Install with: pip install -r requirements.txt

# ── Web Framework ───────────────────────────────────────
fastapi==0.115.0          # High-performance async web framework
uvicorn[standard]==0.30.6 # ASGI server to run FastAPI

# ── LLM (Groq / LLaMA) ─────────────────────────────────
groq==0.11.0              # Groq Python SDK

# ── RAG: Embeddings ─────────────────────────────────────
sentence-transformers==3.2.0  # all-MiniLM-L6-v2 embedding model
torch==2.4.0                  # PyTorch backend for sentence-transformers

# ── RAG: Vector Store ───────────────────────────────────
faiss-cpu==1.8.0              # FAISS vector similarity search (CPU version)
numpy==1.26.4                 # Numerical arrays (used by FAISS + sentence-transformers)

# ── External API Calls ──────────────────────────────────
requests==2.32.3              # HTTP client for weather + currency APIs

# ── Configuration ───────────────────────────────────────
python-dotenv==1.0.1          # Load .env file into os.environ

# ── Data Validation (bundled with FastAPI) ──────────────
pydantic==2.9.2               # Request/response schema validation
